app/requirements.txt
confluent-kafka
psycopg2-binary
tenacity
